DEMO/requirements.txt
Django==1.9
django-tastypie==0.13.3
oauthlib==2.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-dateutil==2.5.3
python-mimeparse==1.5.3
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
